requirement.txt
Django==2.2.6
django-autofixture==0.12.1
django-cors-headers==3.1.1
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
